# booking-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Booking Service
  description: |
    Event booking microservice with seat limiting, waitlist, and concurrency safety.

    ## Booking flow
    1. `POST /bookings` — tries to claim a seat.
       - Seat available → returns `status: confirmed` with a `Booking` object.
       - Event full → caller is added to the waitlist and returns `status: waitlisted`
         with a `WaitlistEntry` object. No error is raised.
    2. `DELETE /bookings/{id}` — cancels a confirmed booking and atomically
       promotes the next waiting user (if any) into that seat.

    ## Concurrency safety
    - Seat counter is kept in Redis (`DECR` is atomic).
    - Cancel + promote runs inside a single PostgreSQL transaction with
      `SELECT … FOR UPDATE SKIP LOCKED` to prevent concurrent double-promotions.
    - Unique DB constraint on `(event_id, user_id)` prevents double-booking at
      the persistence layer.
  version: 1.0.0

servers:
  - url: http://localhost:8082/api/v1

security:
  - ApiKeyAuth: []

paths:
  /bookings:
    post:
      summary: Book an event (confirmed or waitlisted)
      operationId: bookEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
            example:
              event_id: 1
              user_name: "Somchai Jaidee"
              user_email: "somchai@example.com"
              user_phone: "0812345678"
      responses:
        "201":
          description: |
            Booking created. `status` is either `"confirmed"` or `"waitlisted"`.
            Only the matching field (`booking` or `waitlist_entry`) is populated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookEventResult"
              examples:
                confirmed:
                  summary: Seat available — confirmed
                  value:
                    status: "confirmed"
                    booking:
                      id: 42
                      event_id: 1
                      user_id: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
                      status: "confirmed"
                      seat_number: 7
                      created_at: "2026-04-01T10:00:00Z"
                waitlisted:
                  summary: Event full — added to waitlist
                  value:
                    status: "waitlisted"
                    waitlist_entry:
                      id: 5
                      event_id: 1
                      user_id: "b1ffcd88-1a2b-4c3d-9e0f-7a8b9c0d1e2f"
                      position: 3
                      status: "waiting"
                      created_at: "2026-04-01T10:00:01Z"
        "400":
          description: Bad request (missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict — user already has a confirmed booking or is already on the waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}:
    delete:
      summary: Cancel a confirmed booking
      description: |
        Cancels the booking and atomically promotes the next waitlisted user
        into the freed seat (if anyone is waiting). The promotion is invisible
        to the caller; the freed-seat logic is handled server-side.
      operationId: cancelBooking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 42
      responses:
        "200":
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "booking cancelled"
        "400":
          description: Invalid booking ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Booking is already cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/user/{userID}:
    get:
      summary: Get all bookings for a user
      operationId: getUserBookings
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: List of bookings (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/event/{eventID}:
    get:
      summary: Get all bookings for an event
      operationId: getEventBookings
      parameters:
        - name: eventID
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: List of bookings (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    CreateBookingRequest:
      type: object
      required: [event_id, user_name, user_email, user_phone]
      properties:
        event_id:
          type: integer
          format: int64
          example: 1
        user_name:
          type: string
          example: "Somchai Jaidee"
        user_email:
          type: string
          format: email
          example: "somchai@example.com"
        user_phone:
          type: string
          example: "0812345678"

    BookEventResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, waitlisted]
        booking:
          $ref: "#/components/schemas/Booking"
        waitlist_entry:
          $ref: "#/components/schemas/WaitlistEntry"

    Booking:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 42
        event_id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [confirmed, waitlisted, cancelled]
          example: "confirmed"
        seat_number:
          type: integer
          format: int32
          nullable: true
          example: 7
        created_at:
          type: string
          format: date-time
          readOnly: true

    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        event_id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
          example: "Somchai Jaidee"
        user_email:
          type: string
          format: email
        user_phone:
          type: string
        position:
          type: integer
          format: int32
          example: 3
          description: 1-based rank in the waitlist at insertion time
        status:
          type: string
          enum: [waiting, promoted, cancelled]
          example: "waiting"
        created_at:
          type: string
          format: date-time
          readOnly: true

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "already booked this event"